RTIR menu support
If you add the RTIR plugin before the Admin Condititions and Actions
plugin in you RT_SiteConfig.pm, the changes RTIR makes to the menus
breaks the RTIR paths, because the menus added by this plugin will
not find the correct menu location to install them.

This patch adds support for the location that RTIR moves the original
menus to when in the RTIR tree.  RTIR moves the original menu to a
'rtkids' menu item under the top level Menu object.